Fix logging to TensorBoard, presentation of metrics on progress bar and save metrics to YAML file
Previously all tasks and metrics would be on the same graph and the logging would create a nested directory for all tasks and all metrics. Now the directory structure has been fixed, and the task-metrics can be viewed on their own graph. Second, the metrics shown on the progress bar used to be previewed as a per-task dictionary, but now only the metric names are previwed making it more compact. Finally, the metrics are saved to YAML files on a per-task basis.